+++ history.md
@@ -3,6 +3,7 @@
 ## 2.0.4 - pending
 
   * Casting Array properties using anything that inherits from Hash.
+  * Adding `.proxy_method_names` call to help migrations with multiple proxied models.
 
 ## 2.0.3 - 2014-07-04